requirements.txt
google-genai==1.65.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
requests==2.32.5
